+++ dst/app/routes/userPools.js
@@ -26,7 +26,7 @@ const cognitoIdentityServiceProvider = new cinerino.AWS.CognitoIdentityServicePr
     })
 });
 const userPoolsRouter = express_1.Router();
-userPoolsRouter.get('/:userPoolId', permitScopes_1.default([]), rateLimit_1.default, validator_1.default, (req, res, next) => __awaiter(void 0, void 0, void 0, function* () {
+userPoolsRouter.get('/:userPoolId', permitScopes_1.default(['userPools.read']), rateLimit_1.default, validator_1.default, (req, res, next) => __awaiter(void 0, void 0, void 0, function* () {
     try {
         const userPool = yield new Promise((resolve, reject) => {
             cognitoIdentityServiceProvider.describeUserPool({
@@ -52,7 +52,7 @@ userPoolsRouter.get('/:userPoolId', permitScopes_1.default([]), rateLimit_1.defa
         next(error);
     }
 }));
-userPoolsRouter.get('/:userPoolId/clients', permitScopes_1.default([]), rateLimit_1.default, validator_1.default, (req, res, next) => __awaiter(void 0, void 0, void 0, function* () {
+userPoolsRouter.get('/:userPoolId/clients', permitScopes_1.default(['userPools.clients.read']), rateLimit_1.default, validator_1.default, (req, res, next) => __awaiter(void 0, void 0, void 0, function* () {
     try {
         const clients = yield new Promise((resolve, reject) => {
             cognitoIdentityServiceProvider.listUserPoolClients({
@@ -81,7 +81,7 @@ userPoolsRouter.get('/:userPoolId/clients', permitScopes_1.default([]), rateLimi
         next(error);
     }
 }));
-userPoolsRouter.get('/:userPoolId/clients/:clientId', permitScopes_1.default([]), rateLimit_1.default, validator_1.default, (req, res, next) => __awaiter(void 0, void 0, void 0, function* () {
+userPoolsRouter.get('/:userPoolId/clients/:clientId', permitScopes_1.default(['userPools.clients.read']), rateLimit_1.default, validator_1.default, (req, res, next) => __awaiter(void 0, void 0, void 0, function* () {
     try {
         const client = yield new Promise((resolve, reject) => {
             cognitoIdentityServiceProvider.describeUserPoolClient({